Add a cross-arch CI job

+++ src/uv_metadata.c
@@ -101,7 +101,7 @@ static int uvMetadataLoadN(const char *dir,
             }
             return 0;
         }
-        ErrMsgPrintf(errmsg, "%s has size %ju instead of %zu", filename, size,
+        ErrMsgPrintf(errmsg, "%s has size %jd instead of %zu", filename, (intmax_t)size,
                      sizeof content);
         return RAFT_CORRUPT;
     }
